fixed input handler to use correct data type

+++ generator/src/main/java/ca/mcmaster/cas/se2aa4/a2/generator/InputHandler.java
@@ -128,11 +128,11 @@ public class InputHandler {
         }
         if(line.hasOption("pt")) {
             // initialise the member variable
-            polyThick = Integer.parseInt(line.getOptionValue("pt"));
+            polyThick = Float.parseFloat(line.getOptionValue("pt"));
         }
         if(line.hasOption("st")) {
             // initialise the member variable
-            segThick = Integer.parseInt(line.getOptionValue("st"));
+            segThick = Float.parseFloat(line.getOptionValue("st"));
         }
         if(line.hasOption("vt")) {
             // initialise the member variable
